refactor: split dos/machine.rs into machine/ submodules (no behavior change)
machine.rs (2024 lines) -> machine/{mod,vga,vdma,vpic,vpit,vkbd}.rs.
mod.rs keeps PcMachine/MouseState, the I/O port dispatch
(emulate_inb/outb, handle_*_event, queue_irq, pick_pending_vec),
flag consts + vm86 helpers, and re-exports each peripheral via
`pub use` so the super::machine::X surface used by dos.rs / dpmi.rs /
mode_transitions.rs / startup.rs is unchanged. Visibilities widened
to pub(crate)/pub only where a moved item is referenced across the
new module boundary (VirtualPit API, VgaState, VGA_AC_STATE). Pure
mechanical move; //:image builds clean.
