+++ kernel/src/kernel/dos/machine/mod.rs
@@ -1,0 +1,739 @@
+//! PC machine virtualization — the shared "PC age" machine model that both
+//! the DOS and DPMI personalities run on top of.
+//!
+//! This module is policy-free: it owns the per-thread virtual peripherals
+//! (8259 PIC, 8253 PIT, PS/2 keyboard, VGA register set), A20 gate state,
+//! HMA page tracking, and the primitive helpers that decode/execute the
+//! handful of sensitive instructions that trap through the GP fault monitor
+//! (I/O, CLI/STI, IRET, PUSHF/POPF, stack push/pop).
+//!
+//! Personalities call into this module; machine never calls back out.
+//! The two personalities (`vm86` for DOS real-mode, `dpmi` for protected
+//! mode) provide their own GP fault monitors that dispatch software INTs
+//! to their own handlers, but the opcode decode and I/O emulation all
+//! route through the primitives here.
+
+extern crate alloc;
+
+use crate::Regs;
+
+pub const IF_FLAG: u32 = 1 << 9;
+pub const IOPL_MASK: u32 = 3 << 12;
+/// IOPL=1 — kernel-set value for VM86 threads. With IOPL<3 and VME,
+/// CLI/STI/PUSHF/POPF/INT/IRET virtualize through VIF instead of touching
+/// real IF, which is exactly what the cooperative IRQ-injection model
+/// needs. IOPL=0 would also virtualize but trap on a few extras; IOPL=3
+/// would let the guest manipulate real IF and bypass the gate.
+pub const IOPL_VM86: u32 = 1 << 12;
+pub const VM_FLAG: u32 = 1 << 17;
+
+pub const NT_FLAG: u32 = 1 << 14;
+/// Flags that user code cannot change (IOPL, VM, NT)
+pub const PRESERVED_FLAGS: u32 = IOPL_MASK | VM_FLAG | NT_FLAG;
+
+/// HMA spans 16 pages (64KB) starting at page 0x100.
+const HMA_PAGE: usize = 0x100;
+const HMA_PAGE_COUNT: usize = 16;
+/// Shadow region for A20 gate: swap HMA entries here when A20 is off.
+const HMA_SHADOW_PAGE: usize = HMA_PAGE + HMA_PAGE_COUNT;
+
+// ============================================================================
+// VM86 register helpers — 16-bit views of the 32-bit user frame
+// ============================================================================
+
+#[inline]
+pub fn vm86_cs(regs: &Regs) -> u16 {
+    regs.code_seg()
+}
+
+#[inline]
+pub fn vm86_ip(regs: &Regs) -> u16 {
+    regs.ip32() as u16
+}
+
+#[inline]
+pub fn vm86_ss(regs: &Regs) -> u16 {
+    regs.stack_seg()
+}
+
+#[inline]
+pub fn vm86_sp(regs: &Regs) -> u16 {
+    regs.sp32() as u16
+}
+
+#[inline]
+pub fn vm86_flags(regs: &Regs) -> u32 {
+    regs.flags32()
+}
+
+#[inline]
+pub fn set_vm86_cs(regs: &mut Regs, cs: u16) {
+    regs.set_cs32(cs as u32);
+}
+
+#[inline]
+pub fn set_vm86_ip(regs: &mut Regs, ip: u16) {
+    regs.set_ip32(ip as u32);
+}
+
+#[inline]
+pub fn set_vm86_sp(regs: &mut Regs, sp: u16) {
+    let full = (regs.sp32() & 0xFFFF_0000) | sp as u32;
+    regs.set_sp32(full);
+}
+
+#[inline]
+pub fn set_vm86_flags(regs: &mut Regs, flags: u32) {
+    // Merge low 16 bits (user-visible flags), preserve upper EFLAGS (VM, IOPL, VIF, VIP).
+    regs.frame.rflags = (regs.frame.rflags & !0xFFFF) | (flags as u64 & 0xFFFF);
+}
+
+pub(super) mod vga;
+pub use vga::*;
+// ============================================================================
+// PcMachine — per-thread machine state
+// ============================================================================
+pub(super) mod vdma;
+pub(super) use vdma::*;
+/// Policy-free PC machine virtualization — per-thread peripheral state.
+///
+/// Holds the virtual 8259 PIC, 8253 PIT, PS/2 keyboard, VGA register set,
+/// A20 gate state, HMA page tracking, and small latches used by the monitor
+/// decoders (skip_irq, e0_pending).
+///
+/// DOS-specific state (PSP, DTA, heap/free segment, XMS/EMS, FindFirst state,
+/// exec-parent chain) lives directly on `thread::DosState`, not here.
+pub struct PcMachine {
+    pub a20_enabled: bool,
+    pub vpit: VirtualPit,
+    pub vpic: VirtualPic,
+    pub vkbd: VirtualKeyboard,
+    pub mouse: MouseState,
+    pub skip_irq: bool,
+    pub e0_pending: bool,
+    pub vga: VgaState,
+    /// Sound Blaster DMA virtualization (generic virtual 8237 + the
+    /// thread's BLASTER channel/IRQ map). SB itself is passthrough.
+    pub sb: SbDmaState,
+    /// Last value written to CMOS index port 0x70 (NMI bit masked off).
+    /// Reads of port 0x71 pass through to the host CMOS using this index.
+    pub cmos_index: u8,
+
+    /// PM/RM transition state. The pm-side cursor isn't a separate
+    /// field — it's `regs.SS:SP` when user is on pm side, or
+    /// `locked_stack.other_stack` (an `(SS, SP)` pair) when user is on
+    /// rm side. See [`super::mode_transitions::pm_get_stack`].
+    pub locked_stack: super::mode_transitions::LockedStackState,
+}
+
+/// Microsoft Mouse driver (INT 33h) state. Updated by the IRQ 12 packet
+/// stream queued via `queue_irq`; queried by INT 33h subfunctions in dos.rs.
+pub struct MouseState {
+    /// Absolute cursor position, clipped to `[min_x..=max_x]` × `[min_y..=max_y]`.
+    pub x: i16,
+    pub y: i16,
+    /// Current button state. bit 0 = left, 1 = right, 2 = middle.
+    pub buttons: u8,
+    /// User-set clip range. Defaults match a 640×200 mode.
+    pub min_x: i16, pub max_x: i16,
+    pub min_y: i16, pub max_y: i16,
+    /// AX=0Bh delta accumulators since the last read. Reset by `take_motion()`.
+    pub accum_dx: i32,
+    pub accum_dy: i32,
+    /// AX=01/02 hide-show counter. Cursor is visible iff `show_count <= 0`.
+    /// Starts at 1 (hidden); AX=01 decrements, AX=02 increments.
+    pub show_count: i8,
+    /// Text-mode cursor rendering: where (cell offset 0..2000) we currently
+    /// have an inverted attribute, and the original attribute value at that
+    /// cell. `None` means no cursor is currently drawn.
+    pub drawn_at: Option<u16>,
+    pub saved_attr: u8,
+    /// AX=0Ch event handler. CX=mask, ES:DX=handler far address. `mask=0`
+    /// means no handler installed.
+    pub cb_mask: u16,
+    pub cb_seg: u16,
+    pub cb_off: u16,
+    /// Pending event-condition bits since last delivery, plus the deltas
+    /// from the last packet that triggered. Read & cleared by the SLOT_INT74
+    /// dispatcher when it sets up the user-handler call.
+    pub pending_cond: u16,
+    pub last_dx: i16,
+    pub last_dy: i16,
+    /// User GP regs saved across the AX=0Ch handler far-call. ModeSave only
+    /// covers CS/EIP/SS/ESP/EFLAGS/segs; we clobber AX/BX/CX/DX/SI/DI to set
+    /// up the call, so they have to be bracket-saved here and restored by
+    /// the SLOT_INT74_MOUSE_CB_RET slot when the handler RETFs.
+    pub saved_rax: u64,
+    pub saved_rbx: u64,
+    pub saved_rcx: u64,
+    pub saved_rdx: u64,
+    pub saved_rsi: u64,
+    pub saved_rdi: u64,
+}
+
+const VGA_TEXT_BASE: u32 = 0xB8000;
+
+impl MouseState {
+    pub const fn new() -> Self {
+        Self { x: 0, y: 0, buttons: 0,
+               min_x: 0, max_x: 639, min_y: 0, max_y: 199,
+               accum_dx: 0, accum_dy: 0,
+               show_count: 1, drawn_at: None, saved_attr: 0,
+               cb_mask: 0, cb_seg: 0, cb_off: 0,
+               pending_cond: 0, last_dx: 0, last_dy: 0,
+               saved_rax: 0, saved_rbx: 0, saved_rcx: 0,
+               saved_rdx: 0, saved_rsi: 0, saved_rdi: 0 }
+    }
+    /// Apply one PS/2 packet: accumulate raw delta, advance clipped position,
+    /// redraw the cursor at the new cell if it's visible. Returns the
+    /// AX=0Ch condition bits that fired this packet (so the caller can OR
+    /// against the registered event mask and decide to deliver IRQ 12).
+    ///
+    /// AX=0Ch condition bits (from the spec):
+    ///   0x01 = mouse moved
+    ///   0x02 = left button pressed
+    ///   0x04 = left button released
+    ///   0x08 = right button pressed
+    ///   0x10 = right button released
+    ///   0x20 = middle button pressed
+    ///   0x40 = middle button released
+    pub fn apply_packet(&mut self, dx: i16, dy: i16, buttons: u8) -> u16 {
+        self.accum_dx = self.accum_dx.saturating_add(dx as i32);
+        self.accum_dy = self.accum_dy.saturating_add(dy as i32);
+        self.x = (self.x as i32 + dx as i32).clamp(self.min_x as i32, self.max_x as i32) as i16;
+        self.y = (self.y as i32 + dy as i32).clamp(self.min_y as i32, self.max_y as i32) as i16;
+        let prev = self.buttons;
+        let cur = buttons;
+        self.buttons = cur;
+        self.render_if_visible();
+
+        let mut cond: u16 = 0;
+        if dx != 0 || dy != 0 { cond |= 0x01; }
+        let pressed = !prev & cur;
+        let released = prev & !cur;
+        if pressed  & 0x01 != 0 { cond |= 0x02; }
+        if released & 0x01 != 0 { cond |= 0x04; }
+        if pressed  & 0x02 != 0 { cond |= 0x08; }
+        if released & 0x02 != 0 { cond |= 0x10; }
+        if pressed  & 0x04 != 0 { cond |= 0x20; }
+        if released & 0x04 != 0 { cond |= 0x40; }
+
+        // Coalesce condition bits across multiple packets into one delivery;
+        // last_dx/dy reflect only the latest packet (real drivers fire once
+        // per packet, but our delivery is gated on the IRQ-loop tick).
+        self.pending_cond |= cond;
+        self.last_dx = dx;
+        self.last_dy = dy;
+        cond
+    }
+    /// Take and clear the accumulated delta (for INT 33h AX=0Bh).
+    pub fn take_motion(&mut self) -> (i32, i32) {
+        let dx = self.accum_dx;
+        let dy = self.accum_dy;
+        self.accum_dx = 0;
+        self.accum_dy = 0;
+        (dx, dy)
+    }
+
+    /// Text-mode cursor: invert (xor 0x77) the attribute byte at `(x>>3, y>>3)`
+    /// using the standard 8×8 mickey-to-cell ratio. No-op if hidden or
+    /// already drawn at this cell. Real Microsoft Mouse drivers also do this
+    /// in graphics modes via a sprite — we don't (yet); games that go to
+    /// mode 13h hide the driver cursor and draw their own anyway.
+    pub fn render_if_visible(&mut self) {
+        if self.show_count > 0 { return; }
+        let col = (self.x >> 3) as u32;
+        let row = (self.y >> 3) as u32;
+        if col >= 80 || row >= 25 { return; }
+        let offset = (row * 80 + col) as u16;
+        if Some(offset) == self.drawn_at { return; }
+        self.erase_cursor();
+        unsafe {
+            let attr = (VGA_TEXT_BASE + offset as u32 * 2 + 1) as *mut u8;
+            self.saved_attr = core::ptr::read_volatile(attr);
+            core::ptr::write_volatile(attr, self.saved_attr ^ 0x77);
+        }
+        self.drawn_at = Some(offset);
+    }
+
+    /// Restore the original attribute under the current cursor cell.
+    pub fn erase_cursor(&mut self) {
+        if let Some(old) = self.drawn_at.take() {
+            unsafe {
+                let attr = (VGA_TEXT_BASE + old as u32 * 2 + 1) as *mut u8;
+                core::ptr::write_volatile(attr, self.saved_attr);
+            }
+        }
+    }
+
+    /// AX=01h — show cursor: decrement counter; if it just reached 0, draw.
+    pub fn show(&mut self) {
+        self.show_count -= 1;
+        self.render_if_visible();
+    }
+
+    /// AX=02h — hide cursor: increment counter; if it was 0, erase.
+    pub fn hide(&mut self) {
+        if self.show_count <= 0 { self.erase_cursor(); }
+        self.show_count += 1;
+    }
+}
+
+impl PcMachine {
+    pub fn new() -> Self {
+        // A20 starts disabled. HMA_PAGE wraps to user's private low memory
+        // by copying entries[0..16]. HMA_SHADOW_PAGE is left not-present
+        // (arch_user_clean cleared it; map_low_mem_user doesn't touch it),
+        // which is the correct A20-on state when no extended memory is
+        // allocated — set_a20(true) will swap not-present into HMA_PAGE
+        // so HMA accesses fault until XMS maps real extended memory.
+        crate::kernel::startup::arch_copy_page_entries(0, HMA_PAGE, HMA_PAGE_COUNT);
+        Self {
+            a20_enabled: false,
+            vpit: VirtualPit::new(),
+            vpic: VirtualPic::new(),
+            vkbd: VirtualKeyboard::new(),
+            mouse: MouseState::new(),
+            skip_irq: false,
+            e0_pending: false,
+            vga: VgaState::new(),
+            sb: SbDmaState::new(),
+            cmos_index: 0,
+            locked_stack: super::mode_transitions::LockedStackState::new(),
+        }
+    }
+
+    /// Toggle A20 gate: HMA either sees shadow (real content) or wraps to page 0.
+    pub fn set_a20(&mut self, enabled: bool) {
+        if enabled == self.a20_enabled { return; }
+        // Shadow always holds the opposite of what's in HMA.
+        // Swap them to toggle.
+        crate::kernel::startup::arch_swap_page_entries(HMA_SHADOW_PAGE, HMA_PAGE, HMA_PAGE_COUNT);
+        self.a20_enabled = enabled;
+    }
+}
+
+pub(super) mod vpit;
+pub(super) use vpit::*;
+// ============================================================================
+// Virtual hardware — per-thread PIC and keyboard emulation
+// ============================================================================
+
+pub(super) mod vpic;
+pub(super) use vpic::*;
+pub(super) mod vkbd;
+pub(super) use vkbd::*;
+// ============================================================================
+// I/O port emulation
+// ============================================================================
+
+/// Emulate IN from a port using the virtual peripherals.
+pub fn emulate_inb(pc: &mut PcMachine, port: u16) -> u8 {
+    // ISA decodes only A0-A9, so I/O ports alias mod 0x400 (e.g. a
+    // gameport at 0x208 also answers 0x608). DOS-era code relies on
+    // this; the whole DOS I/O surface is <= 0x3FF. Fold the alias
+    // once here so every handler sees the canonical 10-bit address.
+    // Wide-decode I/O (PCI/PnP/ACPI/EISA) is out of scope for the
+    // VM86 guest — we model none of it — and kernel PCI uses a
+    // separate path (arch::inl/outl), not this emulator.
+    let port = port & 0x3FF;
+    match port {
+        // VGA Input Status 1: synthesize retrace signal tied to host wall
+        // time so vsync busy-waits actually wait. Mode 13h refresh ~70 Hz
+        // (14.286 ms/frame), mapped to a 32-step phase:
+        //   phase  0..16 — display active (0x00)            bit0=0 bit3=0
+        //   phase 16..24 — horizontal blank (0x01)          bit0=1 bit3=0
+        //   phase 24..32 — vertical blank   (0x09)          bit0=1 bit3=1
+        // - VL_WaitVBL needs bit3=1 (VBL phase)
+        // - VL_SetScreen needs 6+ consecutive bit0=1/bit3=0 (HBL phase)
+        // - VL_SetCRTC needs bit0=0 (display active phase)
+        // Phase formula: (ticks_ms * 70 cycles/s * 32 phases) / 1000 ms/s,
+        // taken mod 32. With ms-resolution ticks the phase advances ~2.24
+        // steps per host ms, so a tight busy-wait sees the same phase for
+        // many successive reads (gives VL_SetScreen its "consecutive 6+"
+        // criterion for free) but bit transitions happen on real wallclock.
+        0x3DA => {
+            // Read real hardware to reset the AC flip-flop, and track it globally.
+            let _real = crate::arch::inb(0x3DA);
+            unsafe { VGA_AC_STATE.pending_data = false; }
+            let ticks = crate::arch::get_ticks();
+            let phase = ((ticks.wrapping_mul(70 * 32)) / 1000) as u32 & 31;
+            if phase < 16 { 0x00 } else if phase < 24 { 0x01 } else { 0x09 }
+        }
+        // VGA ports — pass through to hardware
+        0x3C0..=0x3D9 | 0x3DB..=0x3DF => crate::arch::inb(port),
+        // Bochs/QEMU VBE Display Interface (BVDI). SeaBIOS uses these
+        // to configure QEMU's emulated VGA, even for legacy modes.
+        // Pass through so SeaBIOS sees real VBE state.
+        0x01CE | 0x01CF | 0x01D0 => crate::arch::inb(port),
+        // Gameport (joystick): we don't model a Sound Blaster or dedicated
+        // gameport card, so on the ISA bus the gameport is unpopulated —
+        // reads return 0xFF (floating data lines, weakly pulled high by
+        // the chipset). Full 0x200-0x20F window: gameport cards decode
+        // loosely and the canonical port is 0x201, but games hit mirrors
+        // across the whole block (and, via 10-bit aliasing folded above,
+        // 0x6xx/0xAxx images of it — e.g. Sokoban polling 0x608-0x60B).
+        // "No card on bus" for the entire window; explicit arm just
+        // suppresses the unhandled-port log during joystick probes.
+        0x200..=0x20F => 0xFF,
+        // Master PIC command (read ISR)
+        0x20 => pc.vpic.isr,
+        // Master PIC data (read IMR)
+        0x21 => pc.vpic.imr,
+        // Keyboard data port — returns current scancode from the virtual 8042.
+        0x60 => pc.vkbd.read_port60(),
+        // Keyboard controller / speaker port used by BIOS IRQ1 acknowledge sequence.
+        0x61 => pc.vkbd.read_port61(),
+        // Keyboard status port (bit 0 = output buffer full)
+        0x64 => if pc.vkbd.has_data() { 1 } else { 0 },
+        0x40 => pc.vpit.read_counter0(),
+        0x41 | 0x42 => 0,
+        // PIT command register not readable
+        0x43 => 0xFF,
+        // CMOS index port: read returns the last index byte (rare).
+        0x70 => pc.cmos_index,
+        // CMOS data port: pass through to host RTC at the saved index.
+        // Host CMOS isn't used by the kernel itself, so this is safe; the
+        // guest sees real time-of-day plus a UIP/VRT-correct status block.
+        0x71 => {
+            crate::arch::outb(0x70, pc.cmos_index);
+            crate::arch::inb(0x71)
+        }
+        // SB DSP/mixer/OPL → straight to the real QEMU sb16/adlib.
+        p if pc.sb.is_passthrough(p) => {
+            let v = crate::arch::inb(p);
+            if p != 0x388 && p != 0x389 {
+                crate::dbg_println!("[SB-IO] in  {:04X} -> {:02X}", p, v);
+            }
+            v
+        }
+        // Virtual 8237 DMA controller. SB channel count register is
+        // served from the interpolated current-count model (drivers
+        // poll it for DMA progress, not just completion).
+        p if Dma8237::owns(p) =>
+            pc.sb.dma_read(p),
+        // Unknown ports: return 0xFF (unpopulated bus). Diagnostic — log to
+        // surface ports the BIOS or guest expects responses on but we don't
+        // virtualize.
+        _ => {
+            crate::dbg_println!("[port] in  {:04X} -> 0xFF (unhandled)", port);
+            0xFF
+        }
+    }
+}
+
+/// Emulate OUT to a port.
+pub fn emulate_outb(pc: &mut PcMachine, port: u16, val: u8) {
+    // ISA 10-bit I/O decode — fold the alias mod 0x400. See `emulate_inb`.
+    let port = port & 0x3FF;
+    match port {
+        // VGA ports — pass through to hardware, track AC flip-flop + index
+        0x3C0 => {
+            unsafe {
+                if !VGA_AC_STATE.pending_data {
+                    VGA_AC_STATE.index = val; // index write — latch full byte (incl. PAS)
+                }
+                VGA_AC_STATE.pending_data = !VGA_AC_STATE.pending_data;
+            }
+            crate::arch::outb(port, val);
+        }
+        0x3C1..=0x3DF => crate::arch::outb(port, val),
+        // Bochs/QEMU VBE Display Interface (BVDI) — see emulate_inb.
+        0x01CE | 0x01CF | 0x01D0 => crate::arch::outb(port, val),
+        // Master PIC command
+        0x20 => {
+            if val == 0x20 {
+                // Non-specific EOI
+                // If keyboard IRQ (bit 1) was in service, and the virtual 8042
+                // still has data ready, IRQ1 should assert again after EOI.
+                let keyboard_in_service = pc.vpic.isr & 0x02 != 0;
+                pc.vpic.eoi();
+                // Real hardware effectively re-asserts IRQ1 if more scancodes are
+                // already visible in the controller when the handler finishes.
+                if keyboard_in_service
+                    && pc.vkbd.has_data()
+                    && !pc.vpic.has_pending_vec(0x09)
+                {
+                    pc.vpic.push(0x09);
+                }
+            }
+        }
+        // Gameport one-shot trigger: no card on this ISA bus window (see
+        // emulate_inb above; full 0x200-0x20F to match the read side and
+        // the 10-bit alias fold). Writes to absent devices are dropped on
+        // real hardware too — silently swallow rather than flood the log.
+        0x200..=0x20F => {}
+        // Master PIC data (write IMR)
+        0x21 => pc.vpic.imr = val,
+        // Slave PIC command / data
+        0xA0 | 0xA1 => {}
+        // Keyboard data port — host-to-device command / parameter byte.
+        // The keyboard's response (ACK, BAT, ID, …) becomes visible at port
+        // 0x60 and asserts IRQ1, exactly as on real hardware.
+        0x60 => {
+            if pc.vkbd.write_port60(val) && !pc.vpic.has_pending_vec(0x09) {
+                pc.vpic.push(0x09);
+            }
+        }
+        // Keyboard controller / speaker port
+        0x61 => pc.vkbd.write_port61(val),
+        // Keyboard controller command
+        0x64 => {}
+        0x43 => pc.vpit.write_command(val),
+        0x40 => pc.vpit.write_counter0(val),
+        0x41 | 0x42 => {}
+        // CMOS index: latch for the next data-port read. Mask off the NMI
+        // disable bit (0x80) — we never want guest writes to toggle host NMI.
+        0x70 => pc.cmos_index = val & 0x7F,
+        // CMOS data writes are dropped — never let the guest mutate host CMOS
+        // (BIOS settings, time-of-day, alarm, etc.).
+        0x71 => {}
+        // SB DSP/mixer/OPL → straight to the real QEMU sb16/adlib.
+        p if pc.sb.is_passthrough(p) => {
+            if p != 0x388 && p != 0x389 {
+                crate::dbg_println!("[SB-IO] out {:04X} <- {:02X}", p, val);
+            }
+            crate::arch::outb(p, val);
+        }
+        // Virtual 8237 DMA controller (generic). After capturing the
+        // write, re-check whether the BLASTER channel just armed and, if
+        // so, remap the guest buffer contiguous + program the real 8237.
+        p if Dma8237::owns(p) => {
+            pc.sb.dma.io_write(p, val);
+            pc.sb.maybe_remap();
+        }
+        // Unknown ports: silently ignore (BIOS probes various ports during mode switches).
+        // Diagnostic — log so we can spot ports SeaBIOS / guest writes to
+        // that we need to virtualize but currently drop.
+        _ => {
+            crate::dbg_println!("[port] out {:04X} <- {:02X} (unhandled)", port, val);
+        }
+    }
+}
+
+// ============================================================================
+// Monitor event handlers — kernel-side completion of I/O bubbled from arch
+// ============================================================================
+
+/// Resolve the linear base of segment `sel`. VM86 uses `sel*16`; PM walks
+/// GDT/LDT via the arch descriptor helpers.
+fn seg_base_for(regs: &Regs, sel: u16) -> u32 {
+    if regs.mode() == crate::UserMode::VM86 {
+        (sel as u32) << 4
+    } else {
+        crate::arch::monitor::seg_base(sel)
+    }
+}
+
+/// Complete an `IN AL/AX/EAX, port` the arch monitor bubbled up. Reads `size`
+/// bytes through `emulate_inb` and writes the result into `regs.rax`.
+pub fn handle_in_event(pc: &mut PcMachine, regs: &mut Regs, port: u16, size: u32) {
+    if size == 2 && matches!(port, 0x01CE | 0x01CF | 0x01D0) {
+        let val = crate::arch::inw(port) as u64;
+        regs.rax = (regs.rax & !0xFFFF) | val;
+        return;
+    }
+
+    let mut val: u64 = 0;
+    for i in 0..size {
+        val |= (emulate_inb(pc, port + i as u16) as u64) << (i * 8);
+    }
+    let mask: u64 = if size >= 4 { 0xFFFF_FFFF } else { (1u64 << (size * 8)) - 1 };
+    regs.rax = (regs.rax & !mask) | (val & mask);
+}
+
+/// Complete an `OUT port, AL/AX/EAX` the arch monitor bubbled up.
+pub fn handle_out_event(pc: &mut PcMachine, regs: &mut Regs, port: u16, size: u32) {
+    let val = regs.rax;
+    if size == 2 && matches!(port, 0x01CE | 0x01CF | 0x01D0) {
+        crate::arch::outw(port, val as u16);
+        return;
+    }
+
+    for i in 0..size {
+        emulate_outb(pc, port + i as u16, (val >> (i * 8)) as u8);
+    }
+}
+
+/// Complete an `INSB/INSW/INSD` (ES:DI ← port, advance DI). Single element —
+/// no REP handling; the CPU traps per iteration when REP is in effect.
+pub fn handle_ins_event(pc: &mut PcMachine, regs: &mut Regs, size: u32) {
+    let port = regs.rdx as u16;
+    let es_base = seg_base_for(regs, regs.es as u16);
+    let di = regs.rdi as u32;
+    for i in 0..size {
+        let b = emulate_inb(pc, port + i as u16);
+        unsafe { *((es_base.wrapping_add(di.wrapping_add(i))) as *mut u8) = b; }
+    }
+    let df = regs.flags32() & (1 << 10) != 0;
+    let delta = if df { (size as u64).wrapping_neg() } else { size as u64 };
+    regs.rdi = regs.rdi.wrapping_add(delta);
+}
+
+/// Complete an `OUTSB/OUTSW/OUTSD` (port ← DS:SI, advance SI). Single element.
+pub fn handle_outs_event(pc: &mut PcMachine, regs: &mut Regs, size: u32) {
+    let port = regs.rdx as u16;
+    let ds_base = seg_base_for(regs, regs.ds as u16);
+    let si = regs.rsi as u32;
+    for i in 0..size {
+        let b = unsafe { *((ds_base.wrapping_add(si.wrapping_add(i))) as *const u8) };
+        emulate_outb(pc, port + i as u16, b);
+    }
+    let df = regs.flags32() & (1 << 10) != 0;
+    let delta = if df { (size as u64).wrapping_neg() } else { size as u64 };
+    regs.rsi = regs.rsi.wrapping_add(delta);
+}
+
+// ============================================================================
+// IRQ delivery — buffer hardware events, drain into the virtual PIC
+// ============================================================================
+
+/// Buffer a hardware event into the virtual PIC / keyboard.
+/// Mode-independent: both VM86 and DPMI share the same virtual devices.
+pub fn queue_irq(pc: &mut PcMachine, event: crate::arch::Irq) {
+    use crate::arch::Irq;
+    match event {
+        Irq::Key(sc) => {
+            let Some(sc) = normalize_scancode(pc, sc) else { return };
+            pc.vkbd.push(sc);
+            if pc.vpic.isr & 0x02 == 0 && !pc.vpic.has_pending_vec(0x09) {
+                pc.vpic.push(0x09);
+            }
+        }
+        Irq::Tick => {
+            let due = pc.vpit.take_pending_irqs();
+            for _ in 0..due {
+                pc.vpic.push(0x08);
+            }
+        }
+        Irq::Mouse { dx, dy, buttons } => {
+            let cond = pc.mouse.apply_packet(dx, dy, buttons);
+            // Raise IRQ 12 (vec 0x74) into the virtual PIC iff the user
+            // registered a handler whose mask intersects this packet's
+            // condition bits. The IVT[0x74] stub at slot SLOT_INT74_MOUSE_CB
+            // will trap to kernel, set up callback args, and far-call the
+            // user handler.
+            if pc.mouse.cb_mask & cond != 0 && !pc.vpic.has_pending_vec(0x74) {
+                pc.vpic.push(0x74);
+            }
+        }
+        Irq::Hw(line) => {
+            // Real SB completion IRQ → relay to the guest vPIC at the
+            // BLASTER-declared vector (IRQ 0-7 → 0x08+line, 8-15 →
+            // 0x70+line-8). Ignore other host lines.
+            if line == pc.sb.irq {
+                // The SB channel's single-cycle transfer just completed:
+                // latch terminal count so a count-register readback
+                // returns 0xFFFF (Dune2's ISR uses this to identify the
+                // IRQ when its software arm flag isn't set).
+                pc.sb.dma.tc[pc.sb.dma8 as usize] = true;
+                pc.sb.dma.tc[pc.sb.dma16 as usize] = true;
+                let vec = if line < 8 { 0x08 + line } else { 0x70 + (line - 8) };
+                if !pc.vpic.has_pending_vec(vec) {
+                    pc.vpic.push(vec);
+                    crate::dbg_println!(
+                        "[SB-DMA] relay SB IRQ{} -> vPIC vec {:#04X} (imr={:#04X} isr={:#04X})",
+                        line, vec, pc.vpic.imr, pc.vpic.isr);
+                }
+                // Ack QEMU-sb16 ourselves (read DSP read-status: 0x22E
+                // 8-bit / 0x22F 16-bit DMA), then re-arm the host line.
+                // This is the host-side device deassert — our concern,
+                // like the host PIC EOI / 8042 read — restoring the
+                // real-hw "device-ack then re-enable" order that the
+                // 8259 in-service bit provides on metal. The guest's
+                // own later 0x22E passthrough read is then a benign
+                // redundant status read. handle_irq still masks the
+                // line on entry; we deassert+unmask here, independent
+                // of if/when the guest ISR runs.
+                crate::arch::inb(pc.sb.io_base + 0x0E);
+                crate::arch::inb(pc.sb.io_base + 0x0F);
+                crate::kernel::startup::arch_rearm_irq(pc.sb.irq);
+            }
+        }
+    }
+}
+
+/// Poll the virtual PIC for a deliverable IRQ, respecting the virtual
+/// interrupt flag and in-service register. Returns the vector to deliver
+/// (and marks it in-service on the master PIC), or `None` if nothing is
+/// ready. The caller is responsible for pushing the interrupt frame
+/// (see `dpmi::reflect_int_to_real_mode` / `dpmi::deliver_pm_irq`).
+pub fn pick_pending_vec(pc: &mut PcMachine, regs: &mut Regs) -> Option<u8> {
+    let vif = regs.frame.rflags & (1u64 << 9) != 0; // IF = virtual interrupt flag
+    if !vif {
+        if pc.vpic.has_pending() {
+            regs.frame.rflags |= 1u64 << 20; // VIP
+        }
+        return None;
+    }
+    if pc.vpic.isr != 0 {
+        return None;
+    }
+    let vec = pc.vpic.pop()?;
+    if vec == 0x09 {
+        if !pc.vkbd.latch() {
+            pc.vpic.push(0x09);
+            return None;
+        }
+    }
+    if vec == 0x0D {
+        crate::dbg_println!(
+            "[SB-DMA] deliver SB vec0D to guest ISR (imr={:#04X} isr={:#04X})",
+            pc.vpic.imr, pc.vpic.isr);
+    }
+    let irq_num = vec.wrapping_sub(8);
+    if irq_num < 8 {
+        pc.vpic.isr |= 1 << irq_num;
+    }
+    // Clear VIP — interrupt is being serviced
+    regs.frame.rflags &= !(1u64 << 20);
+    Some(vec)
+}
+
+/// Read a u16 from a real-mode seg:off address (unaligned-safe, null-safe)
+pub fn read_u16(seg: u32, off: u32) -> u16 {
+    let linear = (seg << 4) + off;
+    let val: u16;
+    unsafe {
+        core::arch::asm!(
+            "movzx {val:e}, word ptr [{addr}]",
+            addr = in(reg) linear,
+            val = out(reg) val,
+            options(readonly, nostack),
+        );
+    }
+    val
+}
+
+/// Write a u16 to a real-mode seg:off address (unaligned-safe, null-safe)
+pub fn write_u16(seg: u32, off: u32, val: u16) {
+    let linear = (seg << 4) + off;
+    unsafe {
+        core::arch::asm!(
+            "mov word ptr [{addr}], {val:x}",
+            addr = in(reg) linear,
+            val = in(reg) val,
+            options(nostack),
+        );
+    }
+}
+
+/// Push a u16 onto the VM86 stack (SS:SP)
+pub fn vm86_push(regs: &mut Regs, val: u16) {
+    let sp = vm86_sp(regs).wrapping_sub(2);
+    set_vm86_sp(regs, sp);
+    write_u16(regs.ss32(), sp as u32, val);
+}
+
+/// Pop a u16 from the VM86 stack (SS:SP)
+pub fn vm86_pop(regs: &mut Regs) -> u16 {
+    let sp = vm86_sp(regs);
+    let val = read_u16(regs.ss32(), sp as u32);
+    set_vm86_sp(regs, sp.wrapping_add(2));
+    val
+}
+
+
+// GP-fault monitor lives in `arch/monitor.rs` now. Kernel only sees the
+// resulting `KernelEvent`s via `do_arch_execute()`; the completion helpers
+// for In/Out/Ins/Outs live at the top of this file (handle_in_event, etc.).
